Resolve GraphQL type mismatch in Cleanup policy type

See merge request https://gitlab.com/gitlab-org/gitlab/-/merge_requests/216609

Merged-by: Dzmitry (Dima) Meshcharakou <12459192-dmeshcharakou@users.noreply.gitlab.com>
Approved-by: Imam Hossain <ihossain@gitlab.com>
Approved-by: Dzmitry (Dima) Meshcharakou <12459192-dmeshcharakou@users.noreply.gitlab.com>

+++ ee/app/graphql/types/virtual_registries/cleanup/policy_last_run_upstream_metrics_type.rb
@@ -14,7 +14,7 @@ module Types
           description: 'Number of entries deleted.',
           experiment: { milestone: '18.7' }
 
-        field :deleted_size, GraphQL::Types::Int, null: false,
+        field :deleted_size, GraphQL::Types::BigInt, null: false,
           hash_key: 'deleted_size',
           description: 'Size in bytes of data deleted.',
           experiment: { milestone: '18.7' }
